engine: Initial size in copy template disk flow
The 'determineTotalImageInitialSize' method is use to compute the initial
size of a disk image based on the source disk image size, including all
the existing snapshots.

This method calls 'getTotalActualSizeOfDisk' which has a logic based on
the source Storage Domain.

In 'ImagesHandler.determineTotalImageInitialSize', the Storage Domain
passed to 'getTotalActualSizeOfDisk' was the destination SD instead of
the source SD.

This fix is relevant for create VM from template (clone)

Change-Id: I61e42bdb8c70de6bdfcb2b431c8c8ca59fbd665e
Bug-Url: https://bugzilla.redhat.com/1438691
Signed-off-by: Fred Rolland <frolland@redhat.com>