webadmin: Inline VolumeListModel#isOptionEnabledOnVolume
This patch inlines VolumeListModel#isOptionEnabledOnVolume and the
predicate it uses as part of the effort of removing Linq#IPredicate
from the codebase.

Besides cleaning up the code and making it easier to maintain, as an
added bonus this patch employs the short-circuit noneMatch method
instead of evaluating the predicate against all the options and only
then checking if it matched none of them, providing a slight
performance improvement.

Change-Id: I3e004b0c2a536ec6520c6b4c64a31d34bedd74a4
Signed-off-by: Allon Mureinik <amureini@redhat.com>