webadmin: RunOnceModel plugged nics checks
This patch replaces the Linq.where call used to check if a VM has
plugged nics with a call to the standard anyMatch.
Besides being more elegant and easier to read, this change introduces
a slight performance improvement by using a short-circuit operation
and avoiding the need to iterate over all the VM's nics and
constructing a collection to temporarily hold them.

Change-Id: Idb52d2e67e57adec5aeb135fdc29d1e7508be455
Signed-off-by: Allon Mureinik <amureini@redhat.com>