webadmin: Adjust DashboardDataServlet query cache handling
Switch dashboard query cache from running at set intervals to waiting for
a set amount of time between the completion of one and the execution of
the next. Additionally, if background cache update is enabled (the default),
cache evistion isn't used. If the cache update is disabled, then cache
eviction is enabled - allowing the queries to be cached for at least a
short time.

Change-Id: I0d4cf7d1289232c7d869e47c72b6709a5caf8815
Bug-Url: https://bugzilla.redhat.com/1438962
Signed-off-by: Scott J Dickerson <sdickers@redhat.com>