core: Inject DAO when needed
Motivation:
The AuditLogableBase has > 60 unneeded protected DAO members to serve all
the command hierarchy. This means that for most of the cases we have a
waste of dozens of injected members that are not in use. More than just
total waist, there is no easy way to understand what are the
dependencies of the command class in question.

Changes:
- Remove ALL the unused DAOs from AuditLogableBase
- Inject private DAOs where needed - no 'protected' usage anymore

Result:
Inject 9-15 DAOs (The ALB has 9 daos injected) instead of steady 66 for
each command.
More readable code as each command states its dependency.

Change-Id: I12e3906a5008ccbe0dcce1cb268dff7a5084789f
Signed-off-by: Roy Golan <rgolan@redhat.com>