restapi: Update to model 4.1.35 and metamodel 1.1.14
The more relevant changes in this new version of the model are the
following:

* Replace generic assigned networks services with services specific to
  the type of object that they are assigned to, in particular data
  centers and clusters.

* Add `driver` attribute to `HostDevice` type.

* Add common concepts document.

* Add appendix containing changes from version 3 to version 4 of
  the API.

* Add `readOnly` attribute to the `DiskAttachment` type.

* Fix the type of the `Host.nics` link. It should be of type
  `HostNic[]`, not `Nic[]`.

The replacement of generic assigned networks service with specific ones
requires a significant refactor of the resources in that area.

The changes in the new version of the metadel affect only the Java SDK,
not the engine, but it is good practice to keep using the latest
version.

Change-Id: Ibef405d6f19330f61cb5ec7193c860557fcfb47b
Related-To: https://bugzilla.redhat.com/1390498
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>