restapi: Update to model 4.2.9 and metamodel 1.2.5
The more relevant changes in this new version of the model are the
following:

* Replace generic assigned networks services with services specific to
  the type of object that they are assigned to, in particular data
  centers and clusters.

* Add `driver` attribute to `HostDevice` type.

* Add specification of input details.

* Add common concepts document.

* Add appendix containing changes from version 3 to version 4 of
  the API.

* Update to metamodel 1.2.5 to handle correctly `or` and `COLLECTION` in
  the specification of input details.

The replacement of generic assigned networks service with specific ones
requires a significant refactor of the resources in that area.

The more relevant change in the new version of the metamodel is the
introduction of constraints for input parameters of methods. This
results in autogenerated code that performs validation and selection of
the correct version of a method according to the input parameters.

Change-Id: Ibef405d6f19330f61cb5ec7193c860557fcfb47b
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>