core: Fix LunDao#removeAll
The removeAll method uses the procedure defined by
setProcedureNameForRemove.
Since it was not called, the removeAll method tried to use the
standard method by naming convention ("deleteluns"), which, of
course, does not exist.

This patch adds the missing call to setProcedureNameForRemove in
LunDaoImpl, a test in LunDaoTest to assure the fix is correct, and
some fixture data to use in this test.

Change-Id: I7bb32324b2412f03d09277c91fd7dd9f9b32729e
Bug-Url: https://bugzilla.redhat.com/1439692
Signed-off-by: Allon Mureinik <amureini@redhat.com>