restapi: Fix adding customized template with VM name
Currently when a template is added from a virtual machine specifying its
name and customizing the disks, there is a NPE when trying to load the
disks, because that part of the code assumes that the id of the VM has
been provided by the user. This patch changes the code so that it uses
the id of the VM retrieved from the database, instead of the id of the
VM provided by the user.

Change-Id: I9e549ca33333c5fab75980f09cec2ad6c835945e
Bug-Url: https://bugzilla.redhat.com/1439970
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>