webadmin: Create KeyWidget
Currently, the KeyValueWidget is being used for the VM select widget
in the New Affinity Group popup dialog. The value component is
completely unused and causes UI issues when the widget is placed in
smaller containers.

This patch creates a KeyWidget, which removes the value component
and changes the KeyValueWidget to inherit from the KeyWidget.

Change-Id: I5038dec6361b4c8f37aa8edc57e66ca9745cff28
Bug-Url: https://bugzilla.redhat.com/1444029
Signed-off-by: Phillip Bailey <phbailey@redhat.com>