webadmin: disable dashboard data if DWH isn't available
Two checks are made to verify DWH configuration. First, the engine DWH
config table is verified. Second, the ability of the JNDI managed DWH
DataSource to create a DB connection is verified. If either test fails,
the query caching and direct queries are disabled.

The change will eliminate error logging for developers who don't need
to work with DWH components.

Bug-Url: https://bugzilla.redhat.com/1441857
Change-Id: I7e42b9e32b32ee75e6d1d290c187d35e1e784926
Signed-off-by: Scott J Dickerson <sdickers@redhat.com>