packaging: setup: Check health
This patch introduces a new utility that can create a standalone script
that can be copied to an older engine machine and ran to get a report.

Currently this report does not do much, but it's expected that we'll
move most of the checks done during engine-setup to be ran also by this
script.

Details:

packaging/setup/bin/bundle-ovirt-engine-health

- creates this "bundle" standalone script

packaging/setup/bin/ovirt-engine-health

- The health checker itself. Can also be ran directly on the engine
  machine, but is mainly intended to be wrapped in the bundle.

packaging/setup/plugins/ovirt-engine-checks:

- A new plugins directory, that is shared by both engine-setup and
ovirt-engine-health.

packaging/setup/plugins/ovirt-engine-health:

- A new plugins directory, specific to the health checker

Also patched the spec and make files to allow building.

Also added packaging/setup/plugins/README to document this directory a
bit.

Also copied the code that checks cluster/DC versions compatibility to a
new plugin in plugins/ovirt-engine-checks, as a POC.

To try this, on the engine machine, after setup, run:

sh /usr/share/ovirt-engine/setup/bin/bundle-ovirt-engine-health

It will create (by default) ovirt-engine-health-check.run in current
directory. Copy this one to the target machine and run it there.

Change-Id: Iafb49d107c3b2b5623d11d0c4bbf907d6398e987
Signed-off-by: Yedidyah Bar David <didi@redhat.com>