core: FenceVolumeJob - avoiding growing number of db records
FenceVolumeJob is executed in order to fence operations that were
submitted and are supposed to be performed on the volume.
In case of failure to fence an operation, the engine may attempt to
fence it again - when the fencing fails constantly the number of
commands will grow till the flow ends. As we store a record for each
command in the commands table that's something should be prevented.
The engine uses the command record just for polling (to avoid executing
another fence operation before the previous one ended) and not for
determining if the fencing succeeded (it polls the entity to verify
that), therefore we are fine with deleting the command entity after the
execution ends.

Change-Id: I62659c3840ce4bc46f7846751d45b779f3606279
Bug-Url: https://bugzilla.redhat.com/1440549
Signed-off-by: Liron Aravot <laravot@redhat.com>