webadmin: Reimplement AbstractUserPortalListModel#onVmAndPoolLoad
This patch reimplements onVmAndPoolLoad using Java 8's stream
capabilities. Using streams allows for a continuous processing of the
data the method retrieves and improves performance by avoiding
allocating temporary lists for each step of the process.

Linq.zip, which is left unused as a consequence of these changes, is
also removed as part of this patch.

Change-Id: Idf62b40e6b528b7c30e222f53b709e2d3d8ba944
Signed-off-by: Allon Mureinik <amureini@redhat.com>