core: running create_functions.sql upon setup
This patch fix an issue when only create_functions.sql was modified (for
example adding a new function)
In the case that this is the only change, the engine-setup will not
recognize the change and the new function will not be created.

The fix adds the create_functions.sql file to the list of files of which
we have a MD5 sum that is used to track changes in thoise files.

Change-Id: I60222b1fb61e2ef76788774cd59d096d4a1b4a31
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1432095
Signed-off-by: emesika <emesika@redhat.com>