webadmin: Don't use Linq#cast in TemplateStorageListModel
Linq#cast redundantly creates a new ArrayList with the same objects
as the source collection has. This is, at best, an O(n) operation and
usually even worse, as the ArrayList's internal array needs to be
reallocated several times.

This patch removes the calls to Linq#cast in TemplateStorageListModel
and replaces them with explicit casts per element of the source
collection where needed.

Change-Id: Id441a97c601229a0abd0a7f61b8d2ff4c66437dc
Signed-off-by: Allon Mureinik <amureini@redhat.com>