webadmin: Remove Linq#cast
Linq#cast redundantly creates a new ArrayList with the same objects
as the source collection has. This is, at best, an O(n) operation and
usually even worse, as the ArrayList's internal array needs to be
reallocated several times.

This patch removes the calls to Linq#cast in TemplateStorageListModel
and replaces them with the proper use of generics where a new
collection is not needed, or with a straight forward call to new
ArrayList<>(source) (which has a fixed O(n) performance guarantee)
where it is.

Note that in many cases the usage of Linq#cast was in a boiler-plate
for-loop that may have been replaced with a more elegant usage of
Java 8's streams, but that's beyond the scope of this patch.

Change-Id: I00b521b803bd379bf7e46afa33b5a652856307cc
Signed-off-by: Allon Mureinik <amureini@redhat.com>