core: remove fence_agents columns from vds view
Fence agents should be dropped from VDS view join.

In case of more then 1 fence agent is defined for a single host the
view returns the same host rows as many times as many agents are defined.
That is bug prone for the view user.
Beside that it has some performance impact on the queries related to VDS view.

This patch removes the fence_agents.* columns from the vds and
vds_with_tags views.

Change-Id: I8abee60df47a155882d7963f85a15c7cd490c870
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1434868
Signed-off-by: emesika <emesika@redhat.com>