core: Free VM lock if attaching a VM from pool failed
If VM is allocated to user from a VM Pool and the VM is not prestarted,
lock is taken on it early to prevent race condition between
AttachUserToVmFromPoolAndRunCommand and VmPoolMonitor. In regular flow,
this lock is freed by RunVmCommand. But if validation of
AttachUserToVmFromPoolAndRunCommand is failed or exception is thrown for
some reason, this lock must be freed explicitly.

Change-Id: I6d13d3196dd7e2557390675066e0b861071d846f
Bug-Url: https://bugzilla.redhat.com/1433924
Bug-Url: https://bugzilla.redhat.com/1432105
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>