core: Block moving host to maintenance if it's running jobs
When a host is running jobs (and specifically storage jobs) we should
block moving it to maintenance until the jobs are completed.
This patch adds support to get the operations running on a host and adds
a validation when it's moved to maintenance.

The following changes are made:
1. Added stored procedure (with tests) that returns steps related to a
subject entity.

2. Add validation in MaintenanceNumberOfVdssCommand.

Regardless of this change, the validation is racey as a job (or a task)
could be started after the validation. Locks should be added in order to
sync between the maintenance flow and the start of jobs on a host.
As this issue existed always, it'll be handled in a separate change -
this patch adds the validation.

Change-Id: Ibfa5947a39844b22bacf28f873bc6842bcf15040
Bug-Url: https://bugzilla.redhat.com/1444982
Signed-off-by: Liron Aravot <laravot@redhat.com>