pm: Wait for host to boot when executing start operation
Previously when start fence action was executed we didn't wait
for the host to boot. This may cause an issue when the host
take too long to boot, so the host monitoring code will be
re-executing the fence operation and the host will never boot.

This patch handle the issue, so we now wait ServerRebootTimeout,
after the 'start' operation is executed and after this timeout
the host status is changed properly.

Change-Id: If872dcab32f8e0f9c2806a64552cf7c6ff28a4b1
Bug-Url: https://bugzilla.redhat.com/1423657
Signed-off-by: Ondra Machacek <omachace@redhat.com>