engine: remove UpdateClusterNetworkClusterValidator
The change enables adding a detached cluster to a DC
Prior the change adding a detached non-empty cluster to a DC, failed
with "changing the management network in a non-empty cluster" message.

On the other hand after adding a detached cluster with hosts to a DC,
the actual network that the engine communicates with the hosts in this
cluster might be different from the declared management network.

Change-Id: I7525fbf7cc7838192df131321bf1935823e3004d
Signed-off-by: Yevgeny Zaspitsky <yzaspits@redhat.com>