core: fix spec_params value in vm_device
The squashing 03_06 patch replaced by mistake the following column
values:
    initrd_url column in vm_static table in file 01100_insert_vm_static.sql
    spec_params column in vm_device table in file 01200_insert_vm_device.sql
    description column in policy_units table in file 00340_insert_policy_units.sql
with UUID while in original it was empty.
As a result an exception is thrown in engine when trying to create a map
from this value for spec_params or the value is incorrect in the other
cases.

This patch fixes that issue.

Change-Id: I281b8953401a9b8e71ee08031af9c9756fc72b32
Signed-off-by: emesika <emesika@redhat.com>