engine: Extract AuditLogable interface from AuditLogableBase
The interface will allow to use the AuditLogable outside of the contexts
of commands safetly. The AuditLogDirector will not make any assumptions
for the implementation of the logged entity.

By defining AuditLogable, we set the ground for introducing a POJO
implementation for AuditLogable which will avoid any need to inject
DAOs. However, it will require to speicfy the data explicitly.

Change-Id: I58ed8ca0ebadb381ecb67531a6c39a80ac90edc4
Signed-off-by: Moti Asayag <masayag@redhat.com>