engine: Introduce AuditLogableImpl
AuditLogableImpl is a POJO which could be used for logging an event by
the AuditLogDirector without the need to consider how the various
members should be evaluated.

It purely logs only the data that was provided by the user, therefore
it is the responsibility of the user to set any relevant data for the event.

As future improvement, we can set the logic to populate the fields which
could be resolved transitively on the DB level, by adding that logic to
the Insert statement of the audit_log table.

Change-Id: I7c2770d31f4ad9dcd39be1700762bf556cd805d1
Signed-off-by: Moti Asayag <masayag@redhat.com>