webadmin: PoolModelBehaviorBaseTest#getBehavior return value
This patch changes the implementation of the
filterClusters(List<Cluster) method of the VmModelBehaviorBase
instance returned by the getBehavior() method to return an empty list
instead of null.

This is done in order to signify to static analysis tools like
Coverity that this method is expected to return a non-null value.

This patch solves Coverity CID 1374427.

Change-Id: I3f791494b1d41c60a407c24c039d0f35d278a45d
Signed-off-by: Allon Mureinik <amureini@redhat.com>