restapi: Add Link-Following support
Add 'link-following' feature. This feature enables API users
to request references inside an entity to be fully fetched
alonside the entity. For example, to recieve a VM along
with its nics:

  GET .../api/vms?follow=nics

Complex usages are possible, with several links/references of
an entity followd, and at several levels of depth, e.g:

  GET .../api/vms?follow=disk_attachments.disk,nics

Bug-Url: https//bugzilla.redhat.com/1463633
Change-Id: I68e242b3668612c70abca9e80d0495b87833ae27
Signed-off-by: Ori Liel <oliel@redhat.com>