benchmarks: Benchmarks various engine API usages
This is an experimental benchmark to measure API requests
to a running engine using a ovirt's java sdk.

Motivation:
Trying to measure throughput and average time of:
- Commands
- Queries

Example benchmarks:
- get api/vms
- get api/hosts
- add external api/events

Additional 'benefits':
using JMH will cause stress on your system, first for warm-ups and later
during measurement - Make sure the benchmarks is a REAL world scenario
as much as possible.

Change-Id: Ibcb8d3e504df7ad3d8fb180c0d8aaa0efce05ff4
Signed-off-by: Roy Golan <rgolan@redhat.com>