core: Adding pretty name to OS version format
This patch adds pretty_name to vds_dynamic table and to the data access
layer.
This field is filled by the VDSM and sent to the engine's VdsBroker
which persist it to the database.
The pretty name is a friendly representation of the OS release/version
and is displayed as a separate field in the UI Host->General->Software
sub tab and is also available from the API when getting the host
information.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1446182
Change-Id: I0b1cc2fa4f9e351eae1b855670b3ffe505a2c618
Signed-off-by: emesika <emesika@redhat.com>
Signed-off-by: Martin Perina <mperina@redhat.com>