core: CreateAllSnapshotsFromVmParameters#disks
Changed disks to a Set<Guid> instead of List<DiskImage>.

While creating a snapshot, the only use for the 'disks' parameter is
for taking the disk IDs. Therefore, constructing a set of disk guids
is more reasonable.

Also, by using guids instead of disk images, checking if a disk
parameter for snapshot is a direct lun would be easier.

In addition, as a part of the patch, the method 'diskImagesNotExist()'
was moved to:
org.ovirt.engine.core.bll.validator.storage.DiskExistenceValidator
and a new test scenario was created for it.

Change-Id: I4b6b6cb38cb28bd42a2a0d81f0ab2d5a2750bddc
Signed-off-by: Shani Leviim <sleviim@redhat.com>