webadmin: AbstractUserPortalListModel#finishSearch streams
Use Java 8's stream to clean up the implementation of finishSearch
and replace the boilerplate for-loop that was used there.

Note that the original implementation had an unwanted(?) side-effect
of sorting the vmOrPoolAndPoolRepresentants passed to it which was
never noticed as it was always called with local variables that were
immediately disposed afterwards. This cleanup also removes that side
effect.

Change-Id: I7628853600dfb35407be0a51edef029065aaeeac
Signed-off-by: Allon Mureinik <amureini@redhat.com>