engine: Use AuditLogDirector directly
The AlertDirecor is a naive wrapper of AuditLogDirector, however, it
doesn't add any value on top of a direct call to AuditLogDirector.

The AlertDirector makes a false assumption that the 'alert' severity is
determined by its logging, however, that is controlled by the severity
of the audit log type.

Therefore, all logging should be done directly by the AuditLogDirector.

Change-Id: Ie2c7cde3822b6fa4b5bf1b7a5c994ef106eeb2ae
Signed-off-by: Moti Asayag <masayag@redhat.com>