frontend: Improve DontPrune aspect code
- added some comments to explain the details
- use Java 8 stream API to reduce code complexity

Note: in aop.xml file, we're configuring AOP weaver to apply
aspects not only to the code we want to advise (3rd party GWT
class ControlFlowAnalyzer) but also to the aspect Java class
itself (DontPrune). Without this, the build fails, so I guess
this is needed simply to ensure the aspect's Java class gets
properly loaded by AspectJ AOP weaver.

Change-Id: I14e84d3193cfcc49582d3cf85e5140be8e364eb0
Signed-off-by: Vojtech Szocs <vszocs@redhat.com>