webadmin: AbstractUserPortalListModel#onVmAndPoolLoad
In the original implementation, for N VMs and M pools this method
filtered the pools by using an O(M*N) nested loop and running over
all the VMs for each pool.
A more efficient approach could be to run over all the VMs and
collect them into a map that counts the number of VMs per pool (an
O(N) operation), and then running over all the pools and checking if
each pool hasn't reached it maximum VMs per user (an O(M) operation),
for a total of O(M+N) time-complexity.

Change-Id: I35ee323e073e2c9dcf68897fdaf745d65334c2b0
Signed-off-by: Allon Mureinik <amureini@redhat.com>