engine: Extract message resolving logic from AuditLogDirector
The message resolving can be extracted into its own class, to be
consumed either by the AuditLogDirector or by other clients which
requires that functionality. There is no good reason to require the
entire AuditLogDirector only for that functionality, and since it is a
specific responsibility, it can be stored on its own class.

This patch perfoms a simple refactor and extracts the methods as static
to the new class. Later-on, we can change the nature of this class to be
injected where needed.

Change-Id: I77f6e268e444ae9941d40f8e07e15bd00b98dd8a
Signed-off-by: Moti Asayag <masayag@redhat.com>