engine: Remove spare injection in GlusterSyncJob
Since all of the data to be logged is already present at the creation
of the audit-log entity, there is no need to inject any beans to the
AuditLogableBase and it can be replaced with a leaner and more
efficient implementation of the AuditLogable.

In addition, the same method to create event in gluster flows can be
reused, therefore introducing and reusing GlusterEventFactory.

Change-Id: I67e7ed9fd59c0bdf168c66ad70e8755a062ecd56
Signed-off-by: Moti Asayag <masayag@redhat.com>