engine: Remove spare injection in PowerSavingBalancePolicyUnit
The information relevant for logging the event is available when
creating the event. Injection can be avoided in this context.
In addition, despite of using ${Host} in the audit log message as the
host name, the event should contain also the host name and host id so
the event can be associated with the actual host in the system.
The same assumption is applied to the cluster in which the action takes
place.

Change-Id: I350047e97335967593490b1756059862d9447585
Signed-off-by: Moti Asayag <masayag@redhat.com>