core: Register OVF_STORE disk at the end of the transaction.
It appears that in some of the postgresql server versions (9.2.18 for
example) a dead lock might occur when attaching a storage domain
that gets upgraded (For example from V3 to V4).

the reason for this dead lock is that AttachStorageDomainToPoolCommand
creates a new transaction and updates the format version
of the storage domain.
As part of this transaction it also calls RegisterDiskCommand
to register the OVF_STORE disks of the storage doman,
RegisterDiskCommand also opens a new transaction and tries
to insert data into image_storage_domain_map table.
Because the new thread in RegisterDiskCommand waits for the lock, which
was aquired as part of the update format, to be freed,
the engine encounters a DB dead lock.

Therefore changing the order of the actions in the trnsaction so the
register of the disks will be before the version update
should avoid this deadlock.

Change-Id: If06cf896c07077e52c71ad33daf1508b3b2c7c86
Bug-Url: https://bugzilla.redhat.com/1446878
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>