engine: fix AddVdsSpmIdCommand to use correct DC id
Before the patch, when the command is run internally, the parent
transaction is suspended (by default), so the vds object, that is
loaded in getExclusiveLocks method and is cached on the command object,
is in the state that ignores the updates that are done in the parent
transaction (command). This way when the command is run from
UpdateClusterCommand and the last has updated an initially detached
cluster to belong to a DC, that new DC id isn't reflected in the vds
object that the command loads from the DB.
The patch fixes the described problem by preventing the suspension of
the parent transaction, so when the command runs internally from another
one it would see the DB data as it was updated by the parent command.

Change-Id: I2ab4e6f1332fe785ad4a19cb8e4dead3264d7815
Signed-off-by: Yevgeny Zaspitsky <yzaspits@redhat.com>