core: ReduceSD - fail when no md devices info
In some user envs the metadata for a block sd was resotored in a way
preventing vdsm to determine the vg metadata device of the domain.
As we don't support that configuration - we'd like to prevent devices
from being reduced from the domain in that case. This patch changes the
reduce command to verify that the metadata devices are present before
allowing to reduce a device from the domain.

Change-Id: Ib4274f3f4a106f1be083bd36dcbe182acc5fe348
Signed-off-by: Liron Aravot <laravot@redhat.com>