core: validate usb unmanaged devices in runVmCommand
In case the Usb policy is set to DISABLED for an "old" VM (<= 3.6)
with an old USB configuration (i.e. more then one USB unmanaged controllers)
and that VM is restarted/rerun/rebooted in engine 4.1:
an error is thrown with host=UNKNOWN (null) since it is thrown in
runVmCommand->initVm()->updateVmDevicesOnRun()  phase while there is still
no host scheduled for this VM.

This scenario is unique since no manual VM configuration update validation is
involved and the user won't know about that prior to restrating/re-running an old VM in
new engine.

To fix that we add the old USB configuration validation to
runVmCommand()->validate() method instead of catching it later on in
initVM() method.

Change-Id: I8876a17d90f5a70bc7bc05814875cbbb40601be3
Bug-Url: https://bugzilla.redhat.com/1435579
Signed-off-by: Sharon Gratch <sgratch@redhat.com>