core: Detach/Remove SD - allowed domain status
This patch prevents detaching/removing a SD while the domain status
isn't Maintenance.
The PreparingForMaintenance status was added in order to make sure
that when a domain moves to Maintenance status it's not accessed by any host
in the data center.

Currently we allow to detach/remove(a local sd) when the domain status
is Inactive, while on that status it still may be accessed by some of the hosts.
This patch changes that and allows to detach/remove a SD only when its on
Maintenance status - an Inactive domain must be deactivated prior to being detached.

Change-Id: I4c18dee175c457aa7aed5ee5bfc8b9803c1aba9d
Bug-Url: https://bugzilla.redhat.com/1418020
Signed-off-by: Liron Aravot <laravot@redhat.com>