core: CreateImagePlaceholder/SyncImageGroupData - setEntityInfo()
The tasks infrastructure uses pre defined prepolling interval in case
the entity info isn't set. While that needs to be changed, this patch
currently sets it in the SyncImageGroupData/CreateImagePlaceholder commands
as well - it'll be removed once we change the infrastructue to not poll based
on that.

This patch is the merge of master commits:
d651f2c8ea
de98b73c87
d1345d9d04

Change-Id: Ifd2b84c2d24dbfbd4efdd9d68292edb2a70e059c
Bug-Url: https://bugzilla.redhat.com/1455469
Signed-off-by: Liron Aravot <laravot@redhat.com>