core: Use DBUnit flat format
This patch converts fixtures.xml from the DBUnit's old XML format to
the newer flat XML format.

The motivation here is maintainability - each row becomes
self-descriptive, and can be read and understood independently,
instead of having to scroll up several (dozens, if not hundreds of)
rows to read the column definition.
Comparing the size of the file before and after this patch could give
a good idea about the improved maintainability. Before the patch,
this file weighed 424KB and contained 10091 rows. After applying this
patch, it weighs 268KB (~37% less) and contains only 743 rows (~93%
less).

The process:
1. The file was converted using DBUnit Transformer [1]
2. The code was changed to load a flat XML instead of the traditional
   XML.
3. Empty rows were added between each table to make the resulting
   file easier to read.
4. FixturesTest, which is no longer needed with the new format, was
   removed.
5. The DAO tests were run to assure they were not broken, naturally.

[1] https://github.com/mureinik/dbunit-transformer

Change-Id: I48c1de98385c39890002ec3d06f7928bab2cb2aa
Signed-off-by: Allon Mureinik <amureini@redhat.com>