core: Upgrade to DBUnit 2.5.3
DBUnit 2.4.8 we used to use is more than six years old(!), and it's
time to retire it.

The newer version improves the native support for Postgres and modern
Java versions, and offers a bunch of bugfixes and performance
enhancements.

Note that DBUnit 2.5.3 is stricter than 2.4.8, and requires some
minor changes to make the test suite pass:
1. The non-default DatabaseConfig.FEATURE_ALLOW_EMPTY_FIELDS property
   needs to be set to true in order to allow <value></value> entries.
2. Timezone and offsets seem to be broken in PostgreSQL. Since these
   values don't seem to affect the tests themselves, the timezones
   and offsets were cleared from the dataset.

Change-Id: I2966ffdfae0cbdd7412f7288ba8217246f9ae0a9
Signed-off-by: Allon Mureinik <amureini@redhat.com>