core: fix watchdog in the generated domain xml
The watchdog device used to be added as unmanaged and unplugged device.
That was alright with the previous mechanism for run VM that used those
watchdog devices anyway, but the new one only uses devices that are
plugged. Logically, the watchdog device is a managed device (set by
the engine).

Change-Id: Ia87c50a95fd7c6d5679093e7c0be98b8386e1056
Signed-off-by: Arik Hadas <ahadas@redhat.com>