packaging: configure auth of ovirt-provider-ovn
Authentication requests to the  ovirt-provider-ovn are
forwarded to engine's SSO. For this reason the host
name and certificate of the engine has to be configured
in the ovirt-provider-ovn.
For authorization the ovirt-provider-ovn requests session
information about the authorization token from the engine's
SSO. To do this, the ovirt-provider-ovn has to authorize
itself as an valid client to the engine's SSO. This is done
by including the combination of client_id and client_secret
in the request for information from the ovirt-provider-ovn
to engine's SSO.
engine-setup generates a new client_secret and registers
the pair of the encoded client_secret and a client_id into
engine's database and stores this pair in the configuration
of ovirt-provider-ovn, too.

Change-Id: Ide901a6dcf8fc5b2a65dceb1edb1d1f6ab11a7c1
Bug-Url: https://bugzilla.redhat.com/1436154
Signed-off-by: Dominik Holler <dholler@redhat.com>
Signed-off-by: Marcin Mirecki <mmirecki@redhat.com>