backend: introduce SyncAttachedDirectLunsCommand
This patch introduces a new command to synchronize direct luns that are
attached to VMs in a given data center.
It is useful when upgrading a data center or when a direct lun
lacks some information because it was created via the REST API without a
host.
The command gets the data center id and synchronizes all the attached
direct luns in it. It can also get one or more attached direct luns ids
and synchronize only them.

Note that the reason that we can't synchronize a floating direct lun is
that it can be done by a host from any of the data centers, and hosts
from different data centers might return different information about the
lun. When a direct lun is attached to a VM, we pick a random host from
the data center that this vm belongs to and use it to synchtonize the
lun's information.

Change-Id: I62b2d6a19400c180ae5869990df92639bbaa9b3b
Related-To:https://bugzilla.redhat.com/1404389
Related-To:https://bugzilla.redhat.com/1455547
Signed-off-by: Idan Shaby <ishaby@redhat.com>