backend: refactor SyncLunsInfoForBlockStorageDomainCommand
The objects updateExistingLuns, saveNewLuns and noOp are consumers that
update/save/do-nothing given a list of luns, and are a part of the
syncing process of storage domains' luns.

This patch changes the way we work with these objects. Instead of
checking their existence in the map that we get from
getLunsToUpdateInDb, each object now contains the relevant info about
itself - requiresDbUpdate and affectsDiscardFunctionality.

This way the process is more general since the interface is decoupled
from the implementation.

Change-Id: I264543c2a9500c51c0263f4fc199b117b2103430
Related-to: https://bugzilla.redhat.com/1439692
Signed-off-by: Idan Shaby <ishaby@redhat.com>