backend: remove old luns from db on sd sync
This patch removes the method cleanupLunsFromDb and instead wraps its
logic inside a LunUpdateHandler object, just like when updating or
saving new luns.

It also fixes a bug where luns are being removed only if new or existing
and not up to date luns were found.
That is, we now always check if old luns exist and then remove them.

Change-Id: Ia95a44ef269a5daecf71a2bddfa813919d179ca6
Bug-Url: https://bugzilla.redhat.com/1439692
Signed-off-by: Idan Shaby <ishaby@redhat.com>