core: fix free space validation when performing merge
When performing a merge operation on snapshot consisting
of multiple disks on the same storage domain, the operation
will fail because of IllegalStateException caused by duplicate
key, because a separate entry is created for each disk and because
they are on the same storage domain, the map will have the same
key for multiple entries.

This is solved by mapping a storage domain (guid) to a list of
SubchainInfo instead, this way all subchains from the same domain
will be under the same entry.

Change-Id: Ie60744140d84f979732e352559d49a1c1cfbf143
Bug-Url: https://bugzilla.redhat.com/1450866
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>