core: handle 'migration failure due to maintenance' logging
In case a vm migration is running due to vds in
VDSStatus.PreparingForMaintenance status (i.e. vds moved to maintenance and
caused a vm migration to start), then in case the migration failed, we
need to check if there is a destination host existed or not.

Before this fix, the audit log message in case of no destination vds
left to re-run on, included an "UNKNOWN" destination vds name:
"Migration failed  while Host is in 'preparing for maintenance' state.
Consider manual intervention: stopping/migrating Vms as Host's state will not
turn to maintenance while VMs are still running on it.(VM: STAYING_ALIVE-8,
Source: dell-r210ii-13, Destination: <UNKNOWN>)"

In this patch we added a new message for omitting destination
host if none exist after re-run trials.

Change-Id: I74605aa7d37551169911e4a406dac094e65e0566
Bug-Url: https://bugzilla.redhat.com/1436397
Signed-off-by: Sharon Gratch <sgratch@redhat.com>